third_party/blink/web_tests/platform/generic/http/tests/security/isolatedWorld/dispatchEvent-expected.txt

Tests that event handler is called in correct world.
Sending from world 0...
In main world: Main World
in isoloated world 0: Isolated World 0
Sending from world 1...
In main world: Main World
in isoloated world 0: Isolated World 0
Sending from main world...
In main world: Main World
in isoloated world 0: Isolated World 0
